+++ api/core/src/modules/knowledge/knowledge.module.ts
@@ -1,10 +1,11 @@
 import { Module } from "@nestjs/common";
 import { KnowledgeController } from "./knowledge.controller";
+import { KnowledgeRepository } from "./knowledge.repository";
 import { KnowledgeService } from "./knowledge.service";
 
 @Module({
   controllers: [KnowledgeController],
-  providers: [KnowledgeService],
+  providers: [KnowledgeRepository, KnowledgeService],
   exports: [KnowledgeService]
 })
 export class KnowledgeModule {}